requirements-backend.txt
# Acty — backend Python dependencies
# Install: pip install -r requirements.txt

# ── Core API ──────────────────────────────────────────────────────────────────
fastapi>=0.100.0
uvicorn>=0.28.0
python-dotenv>=1.0.0
httpx>=0.25.0
asyncpg>=0.30.0
ollama>=0.1.0
python-multipart>=0.0.6

# ── Data / ML ─────────────────────────────────────────────────────────────────
numpy>=1.24.0
pandas>=2.0.0
scikit-learn>=1.3.0
xgboost>=2.0.0
torch>=2.0.0

# ── RAG pipeline ─────────────────────────────────────────────────────────────
sentence-transformers>=2.6.0
chromadb>=0.4.0
pymupdf>=1.23.0
pdfplumber>=0.10.0
pytesseract>=0.3.10
Pillow>=9.5.0

# ── Privacy / federated learning ─────────────────────────────────────────────
flwr>=1.7.0
opendp>=0.9.0

# ── BYOK multi-provider LLM SDKs ─────────────────────────────────────────────
anthropic>=0.35.0
openai>=1.40.0
google-genai>=1.0.0
cohere>=4.0.0
mistralai>=0.1.0
groq>=0.9.0

# ── Key encryption ────────────────────────────────────────────────────────────
cryptography>=41.0.0
python-jose[cryptography]>=3.3.0

# ── Database ──────────────────────────────────────────────────────────────────
psycopg2-binary>=2.9.0
alembic>=1.12.0
sqlalchemy>=2.0.0

# ── Testing & Development (install only in dev) ───────────────────────────────
# pytest>=7.0.0
# pytest-asyncio>=0.21.0
# black>=23.0.0
# flake8>=6.0.0
